requirements.txt
AccessControl==4.2
Acquisition==4.6
aiodns==2.0.0
aiohttp==3.8.0
aiosignal==1.2.0
apipkg==1.5
appdirs==1.4.4
APScheduler==3.8.0
aspy.refactor-imports==2.1.1
astroid==2.8.4
asttokens==2.0.4
async-timeout==4.0.0
attrdict==2.0.1
attrs==21.2.0
AuthEncoding==4.1
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.7
backcall==0.2.0
base58==2.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bitarray==1.2.2
black==21.9b0
bleach==4.0.0
blinker==1.4
bpyutils==0.2.0
bs4==0.0.1
BTrees==4.7.2
cached-property==1.5.1
certifi==2021.10.8
cffi==1.14.0
cfgv==3.1.0
Chameleon==3.8.1
chardet==4.0.0
charset-normalizer==2.0.7
chromedriver==2.24.1
click==8.0.3
colorama==0.4.3
colored==1.4.2
colored-traceback==0.3.0
coloredlogs==14.0
commonmark==0.9.1
configobj==5.0.6
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.2.1
cycler==0.10.0
cytoolz==0.11.0
daemon==1.2
daemons==1.3.2
dataclassy==0.11.0
dateparser==1.0.0
DateTime==4.3
dbfread==2.0.4
dbus-python==1.2.16
decorator==4.4.2
dill==0.3.4
discord.py==1.7.3
distlib==0.3.1
distro==1.4.0
docopt==0.6.2
docutils==0.16
eip712==0.1.0
elpy==1.999
entrypoints==0.3
epc==0.0.5
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.1
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
ExtensionClass==4.5.1
ezenv==0.91
fasteners==0.14.1
filelock==3.0.12
flake8==4.0.1
flake8-mypy==17.8.0
flake8-noqa==1.2.0
Flask==1.1.2
forex-python==1.5
frozenlist==1.2.0
gdown==3.12.2
git-filter-repo==2.32.0
gitdb==4.0.2
GitPython==3.1.0
grip==4.5.2
gshell==5.5.2
gunicorn==20.0.4
hexbytes==0.2.2
httplib2==0.14.0
humanfriendly==8.1
hyperlink==20.0.1
hypothesis==6.23.2
identify==1.4.19
idna==3.2
ijson==2.6.1
imageio==2.9.0
importchecker==2.0
importlib-metadata==3.10.0
importlib-resources==5.1.2
importmagic==0.1.7
incremental==17.5.0
inflection==0.5.0
iniconfig==1.1.1
install==1.3.4
ipdb==0.13.3
ipfshttpclient @ https://github.com/avatar-lavventura/py-ipfs-http-client/archive/refs/heads/master.zip
ipython==7.27.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
jeepney==0.6.0
Jinja2==2.11.2
jsonpatch==1.22
jsonpickle==2.0.0
jsonpointer==2.0
jsonschema==3.2.0
keyring==21.2.1
kiwisolver==1.3.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lmdb==1.2.0
lockfile==0.12.2
louis==1.3
lru-dict==1.1.7
lxml==4.6.3
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.3.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mdv==1.7.4
mercurial==5.4.2
meza==0.42.5
monotonic==1.6
more-itertools==8.8.0
morphys==1.0
multiaddr==0.0.9
multidict==5.2.0
multihash==0.1.1
MultiMapping==4.1
multipart==0.2.4
multitasking==0.0.9
mypy==0.910
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
netifaces==0.10.4
NeuroTools==0.3.1
ng==1.2.0
nodeenv==1.6.0
numpy==1.19.4
oauthlib==3.1.0
packaging==21.0
pandas==1.3.1
paramiko==2.7.1
parsimonious==0.8.1
parso==0.7.1
PasteDeploy==2.1.0
path-and-address==2.0.1
pathspec==0.9.0
pbr==5.5.1
pep517==0.10.0
Persistence==3.1
persistent==4.6.4
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
Pillow==8.0.1
pipenv==2020.11.15
pipupgrade==1.10.1
pkginfo==1.5.0.1
platformdirs==2.4.0
pluggy==1.0.0
ppi==0.0.4
pre-commit==2.5.1
protobuf==3.18.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py-cid==0.3.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycares==4.0.0
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.11.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyGObject==3.36.0
pygogo==0.13.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.11.1
pymongo==3.11.0
pymy==0.0.8
PyMySQL==0.9.3
PyNaCl==1.4.0
pyocclient==0.6
pyOpenSSL==20.0.1
pypandoc==1.6.3
pyparsing==2.4.7
pyright==0.0.9
pyrsistent==0.18.0
pyserial==3.4
PySimpleGUI==4.33.0
PySocks==1.7.1
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-baseconv==1.2.2
python-binance==1.0.11
python-daemon==2.3.0
python-dateutil==2.8.1
python-debian==0.1.39
python-dotenv==0.16.0
python-gettext==4.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.1
python-slugify==1.2.6
pythx==1.6.1
pytz==2020.1
pyxdg==0.26
PyYAML==5.4.1
readme-renderer==26.0
regex==2021.10.8
reportlab==3.5.67
requests==2.26.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
RestrictedPython==5.0
rfc3986==1.4.0
rich==10.12.0
rlp==1.2.0
roman==3.3
rope==0.19.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
seed-isort-config==2.2.0
selenium==3.141.0
selinux==0.2.1
semantic-version==2.8.5
service-identity==18.1.0
sexpdata==0.0.3
simplejson==3.16.0
six==1.16.0
SLIP==0.3.8
smmap==3.0.1
snowballstemmer==2.1.0
sortedcontainers==2.4.0
soupsieve==2.2.1
ssh-import-id==5.10
sshtunnel==0.1.4
stevedore==3.3.0
sudo==1.0.0
tabulate==0.8.9
termcolor==1.1.0
testresources==2.0.1
tk==0.1.0
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
tox==3.24.3
tqdm==4.62.3
traitlets==4.3.3
transaction==3.0.0
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.3
types-filelock==0.1.5
types-pytz==2021.1.2
types-PyYAML==6.0.0
types-termcolor==1.1.1
typing-extensions==3.10.0.2
tzlocal==2.1
ujson==4.0.2
Unidecode==1.2.0
urllib3==1.26.7
varint==1.0.2
virtualenv==20.0.10
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
vvm==0.1.0
vyper==0.3.0
wadllib==1.3.3
waitress==1.4.4
wcwidth==0.2.5
web3==5.24.0
webdriver-manager==3.4.2
webencodings==0.5.1
WebOb==1.8.7
websockets==9.1
WebTest==2.0.35
Werkzeug==1.0.1
wrapt==1.13.1
WSGIProxy2==0.4.6
WTForms==2.3.1
xlrd==1.2.0
yapf==0.31.0
yarl==1.7.2
yfinance==0.1.63
z3c.pt==3.3.0
zc.lockfile==2.0
ZConfig==3.5.0
zExceptions==4.1
zipp==3.4.1
ZODB==5.6.0
zodbpickle==2.0.0
zope.browser==2.3
zope.component==4.6.2
zope.configuration==4.4.0
zope.contentprovider==4.2.1
zope.contenttype==4.5.0
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.exceptions==4.4
zope.hookable==5.1.0
zope.i18n==4.7.0
zope.i18nmessageid==5.0.1
zope.interface==5.4.0
zope.location==4.2
zope.proxy==4.4.0
zope.publisher==6.0.2
zope.schema==6.1.0
zope.security==5.1.1
zope.tales==5.1
zope.testing==4.9
zope.traversing==4.4.1
vulcano==0.5.3
